authMachine => createAuthenticatorMachine
