+++ packages/ui/src/machines/authenticator/actors/index.ts
@@ -1,0 +1,4 @@
+export * from './signIn';
+export * from './signUp';
+export * from './signOut';
+export * from './resetPassword';
