+++ packages/ui/src/machines/index.ts
@@ -1,1 +1,1 @@
-export * from './authMachine';
+export * from './authenticator';
